+++ src/UnitTestEx.NUnit/Internal/GenericTesterT.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/UnitTestEx
 
+using CoreEx.Hosting;
 using UnitTestEx.Generic;
 
 namespace UnitTestEx.NUnit.Internal
@@ -7,7 +8,7 @@ namespace UnitTestEx.NUnit.Internal
     /// <summary>
     /// Provides the <b>NUnit</b> generic testing capability.
     /// </summary>
-    public class GenericTester : GenericTesterBase<GenericTester>
+    public class GenericTester<TEntryPoint> : GenericTesterBase<TEntryPoint, GenericTester<TEntryPoint>> where TEntryPoint : IHostStartup, new()
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="GenericTester"/> class.