+++ src/UnitTestEx.Xunit/Internal/ValidationTesterT.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/UnitTestEx
 
+using CoreEx.Hosting;
 using CoreEx.Validation;
 using UnitTestEx.Generic;
 using Xunit.Abstractions;
@@ -9,10 +10,10 @@ namespace UnitTestEx.Xunit.Internal
     /// <summary>
     /// Provides the <b>Xunit</b> <see cref="IValidator"/> testing capability.
     /// </summary>
-    public class ValidationTester : ValidationTesterBase<ValidationTester>
+    public class ValidationTester<TEntryPoint> : ValidationTesterBase<TEntryPoint, ValidationTester<TEntryPoint>> where TEntryPoint : IHostStartup, new()
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="ValidationTester"/> class.
+        /// Initializes a new instance of the <see cref="ValidationTester{TEntryPoint}"/> class.
         /// </summary>
         /// <param name="output">The <see cref="ITestOutputHelper"/>.</param>
         internal ValidationTester(ITestOutputHelper output) : base(new XunitTestImplementor(output)) { }