+++ tests/UnitTestEx.NUnit.Test/Other/GenericTest.cs
@@ -1,4 +1,7 @@
-﻿using NUnit.Framework;
+﻿using Microsoft.Extensions.DependencyInjection;
+using NUnit.Framework;
+using System;
+using System.Threading.Tasks;
 using UnitTestEx.Expectations;
 
 namespace UnitTestEx.NUnit.Test.Other
@@ -22,5 +25,42 @@ namespace UnitTestEx.NUnit.Test.Other
             test.ExpectErrorType(CoreEx.Abstractions.ErrorType.ValidationError, "Badness.")
                 .Run(() => throw new CoreEx.ValidationException("Badness."));
         }
+
+        [Test]
+        public void Run_Service()
+        {
+            using var test = GenericTester.Create().ConfigureServices(services => services.AddSingleton<Gin>());
+
+            test.Run<Gin, int>(gin => gin.Pour())
+                .AssertSuccess()
+                .Assert(1);
+
+            test.Run<Gin, int>(gin => gin.PourAsync())
+                .AssertSuccess()
+                .Assert(1);
+
+            test.Run<Gin>(gin => gin.Shake())
+                .AssertSuccess();
+
+            test.Run<Gin>(gin => gin.ShakeAsync())
+                .AssertSuccess();
+
+            test.Run<Gin>(gin => gin.Stir())
+                .AssertException<DivideByZeroException>("As required by Bond; shaken, not stirred.");
+
+            test.Run<Gin>(gin => gin.StirAsync())
+                .AssertException<DivideByZeroException>("As required by Bond; shaken, not stirred.");
+        }
+    }
+
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Just for testing aye bro!")]
+    public class Gin
+    {
+        public void Stir() => throw new DivideByZeroException("As required by Bond; shaken, not stirred.");
+        public Task StirAsync() => throw new DivideByZeroException("As required by Bond; shaken, not stirred.");
+        public void Shake() { }
+        public Task ShakeAsync() => Task.CompletedTask;
+        public int Pour() => 1;
+        public Task<int> PourAsync() => Task.FromResult(1);
     }
 }